create helper functions
