+++ client/src/components/Modal/RenameCategory.vue
@@ -38,6 +38,7 @@
 
 <script>
 import axios from 'axios';
+import helper from '../../services/helper.js';
 export default {
     name: 'DeleteCategory',
     data() {
@@ -49,7 +50,7 @@ export default {
     created: function() {
         axios.defaults.headers.common['Authorization'] = `Bearer ${this.$store.auth.token}`;
         //clone the selected feed from the store
-        this.index = this.findIndexById(this.$store.data.categories, this.$store.data.currentSelection.categoryId);
+        this.index = helper.findIndexById(this.$store.data.categories, this.$store.data.currentSelection.categoryId);
         this.category = this.$store.data.categories[this.index];
     },
     methods: {
@@ -74,17 +75,6 @@ export default {
                     this.$store.data.setShowModal('')
                 }
             );
-        },
-        // This function finds the index of an object in an array by its id
-        findIndexById: function(array, id) {
-            var index = -1
-            for(var i = 0; i < array.length; i++) {
-                if(array[i].id == id) {
-                    index = i;
-                    break;
-                }
-            }
-            return index;
         }
     }
 }
